contract change JS into TS
