+++ .husky/.gitignore
@@ -1,1 +1,1 @@
-
+_
