+++ package.json
@@ -15,8 +15,10 @@
     "prepare": "husky install"
   },
   "devDependencies": {
+    "@types/file-saver": "^2.0.2",
     "@types/lodash": "^4.14.168",
     "@types/qs": "^6.9.6",
+    "@types/request": "^2.48.5",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^4.0.0",
     "@umijs/fabric": "^2.2.2",
@@ -32,13 +34,13 @@
     "eslint-plugin-react": "^7.20.3",
     "eslint-plugin-react-hooks": "^4.0.8",
     "file-loader": "^6.2.0",
+    "husky": "^6.0.0",
     "jest": "^26.6.3",
     "lint-staged": "^10.0.7",
     "prettier": "^2.2.0",
     "terser-webpack-plugin": "^5.1.1",
     "umi": "^3",
-    "url-loader": "^4.1.1",
-    "husky": "^6.0.0"
+    "url-loader": "^4.1.1"
   },
   "gitHooks": {
     "pre-commit": "lint-staged"
@@ -59,7 +61,7 @@
     "@mdi/font": "^4.8.95",
     "@mdi/js": "^4.8.95",
     "@mdi/react": "^1.2.1",
-    "antd": "4.10.1",
+    "antd": "^4.15.5",
     "echarts": "^4.9.0",
     "file-saver": "^2.0.5",
     "less": "^3.9.0",
