+++ src/pages/about/contract/myContract/_config.ts
