+++ src/pages/about/contract/privacyStrategy/components/ConfigStrategy/index.tsx
@@ -2,9 +2,29 @@ import React, { useEffect, useState } from 'react';
 import { connect } from 'dva';
 import { Checkbox, Button, Modal } from 'antd';
 import style from './index.less';
+import { ConnectState } from '@/models/connect';
+import { Dispatch, Effect } from 'umi';
 
-function ConfigStrategy(props) {
-  const { User, visible, editParams = {}, onCancel, getPageListOPrivacyStrategy, dispatch, configLoading = false } = props;
+export interface ConfigStrategyProps {
+  User: ConnectState['User'];
+  visible: boolean;
+  editParams: any;
+  onCancel: any;
+  getPageListOPrivacyStrategy: Effect;
+  dispatch: Dispatch;
+  configLoading: boolean;
+  PrivacyStrategy: ConnectState['PrivacyStrategy'];
+}
+function ConfigStrategy(props: ConfigStrategyProps) {
+  const {
+    User,
+    visible,
+    editParams = {},
+    onCancel,
+    getPageListOPrivacyStrategy,
+    dispatch,
+    configLoading = false,
+  } = props;
   const { networkName, leagueName } = User;
   const [initValue, setInitValue] = useState([]);
   const [memberList, setMemberList] = useState([]);
@@ -43,12 +63,13 @@ function ConfigStrategy(props) {
     ],
   };
 
-  const onChange = (checkedValues) => {
+  const onChange = (checkedValues: any) => {
     setInitValue(checkedValues);
   };
 
+  //TODO:strategyMemberList来源于model.model里是空数组，so,item里面没有memberName属性
   const getValue = () => {
-    const value = [];
+    let value = [];
     strategyMemberList.forEach((item) => {
       value.push(item.memberName);
     });
@@ -56,7 +77,7 @@ function ConfigStrategy(props) {
   };
 
   const getInitValue = () => {
-    const value = [];
+    let value = [];
     strategyMemberList.forEach((item) => {
       if (item.checked) {
         value.push(item.memberName);
@@ -89,7 +110,7 @@ function ConfigStrategy(props) {
   );
 }
 
-export default connect(({ User, PrivacyStrategy, loading }) => ({
+export default connect(({ User, PrivacyStrategy, loading }: ConnectState) => ({
   User,
   PrivacyStrategy,
   configLoading: loading.effects['PrivacyStrategy/updateStrategyMember'],
